Возвращение удалённого обновления 0.1

+++ TrainWindowsFormsApp/TrainWindowsFormsApp/Exercise.cs
@@ -7,12 +7,12 @@ namespace TrainWindowsFormsApp
         // Тренировки выстроены на периодичность: сила, виносливость(многоповторность) и по принципу Табата (20сек. нагрузки, 10сек. отдыха)
         public string Name;                 // Название упражнения
         public List<Dictionary<string, string>> typesTrainingList = new List<Dictionary<string, string>>();  // Список, содержащий словари с типами тренировк
-        public int MaxRepeat;               // Максимально количество повторений, при достижении которого следует указать новую нагрузку
+        public float MaxRepeat;               // Максимально количество повторений, при достижении которого следует указать новую нагрузку
         public string Remark;               // Примечания
 
         public Exercise(string name,
                         string strengthRepeat, string staminaRepeat,
-                        int maxRepeat,
+                        float maxRepeat,
                         string strengthLoad, string staminaLoad, string tabataLoad, string intervalLoad,
                         string intervalExercise,
                         string remark = null)
